Estimate incidence from cross-sectional data

Estimate incidence from cross-sectional data

est.inc(
  X,
  N0,
  mu_estimate,
  mu_CI_low,
  mu_CI_high,
  mu_CI_level = 0.95,
  mu_prior = fit_gamma_distribution(mean = mu_estimate, lower = mu_CI_low, upper =
    mu_CI_high, confidence_level = mu_CI_level),
  mu_shape = mu_prior[1],
  mu_scale = mu_prior[2],
  incidence_CI_level = 0.95,
  alpha = 1 - incidence_CI_level,
  tol = 1e-08,
  max_incidence_CI_range = c(0, 100)
)

Arguments

X
The number of MAA-positive participants in the cross-sectional sample

N0
The number of seronegative participatns in the cross-sectional sample

mu_estimate
The estimated mean window period, mu

mu_CI_low
The lower confidence limit for the mean window period, mu (can skip this parameter if providing mu_shape and mu_scale)

mu_CI_high
The upper confidence limit for the mean window period, mu (can skip this parameter if providing mu_shape and mu_scale)

mu_CI_level
The confidence level for the mean window period, mu (can skip this parameter if providing `mu_prior`, or `mu_shape` and `mu_scale`)

mu_prior
The Gamma parameters for the assumed prior distribution on mu (can skip this argument if providing mu_shape and mu_scale directly)

mu_shape
The shape (gamma) parameter for the prior distribution of the mean window period, mu.

mu_scale
The scale (beta) parameter for the prior distribution of the mean window period, mu.

incidence_CI_level
The desired confidence interval coverage probability

alpha
One minus the desired coverage probability

tol
The desired tolerance for matching the specified confidence level

max_incidence_CI_range
The range of incidence rate values to search over (if not wide enough, an error will be produced).

Value

a numeric vector containing the estimated incidence rate, lower confidence limit, and upper confidence limit, in units of new infections per 100 person-years.